@@ -53,7 +53,7 @@ type installCommand struct {
 	cli cli.CLI
 }
 
-type installOptions struct {
+type InstallOptions struct {
 	releaseName             string
 	canaryOperatorNamespace string
 	istioNamespace          string
@@ -62,11 +62,13 @@ type installOptions struct {
 	DumpResources bool
 }
 
-func NewInstallOptions() *installOptions {
-	return &installOptions{}
+// NewInstallOptions get InstallOptions
+func NewInstallOptions() *InstallOptions {
+	return &InstallOptions{}
 }
 
-func NewInstallCommand(cli cli.CLI, options *installOptions) *cobra.Command {
+// NewInstallCommand get installCommand
+func NewInstallCommand(cli cli.CLI, options *InstallOptions) *cobra.Command {
 	c := &installCommand{
 		cli: cli,
 	}
@@ -103,7 +105,7 @@ It can only dump the applicable resources with the '--dump-resources' option.
 	return cmd
 }
 
-func (c *installCommand) run(cli cli.CLI, options *installOptions) error {
+func (c *installCommand) run(cli cli.CLI, options *InstallOptions) error {
 	err := c.validate(options.istioNamespace)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, istioNotFoundErrorTemplate, err)
@@ -198,7 +200,7 @@ func (c *installCommand) validate(istioNamespace string) error {
 	}
 
 	if len(pods.Items) > 0 {
-		errors.Errorf("Istio sidecar injector not healthy yet in '%s'", istioNamespace)
+		return errors.Errorf("Istio sidecar injector not healthy yet in '%s'", istioNamespace)
 	}
 
 	return errors.Errorf("could not find Istio sidecar injector in '%s'", istioNamespace)
