@@ -96,10 +96,10 @@ type Values struct {
 	} `json:"prometheus"`
 
 	Grafana struct {
-		Enabled         bool                        `json:"enabled"`
-		Image           helm.Image                  `json:"image"`
-		Resources       corev1.ResourceRequirements `json:"resources,omitempty"`
-		Security struct {
+		Enabled   bool                        `json:"enabled"`
+		Image     helm.Image                  `json:"image"`
+		Resources corev1.ResourceRequirements `json:"resources,omitempty"`
+		Security  struct {
 			Enabled       bool   `json:"enabled"`
 			UsernameKey   string `json:"usernameKey,omitempty"`
 			SecretName    string `json:"secretName,omitempty"`
@@ -115,16 +115,16 @@ type Values struct {
 	} `json:"ingressgateway"`
 
 	AuditSink struct {
-		Enabled bool `json:"enabled"`
-		Image helm.Image `json:"image"`
-		Resources corev1.ResourceRequirements `json:"resources"`
-		Tolerations []corev1.Toleration `json:"tolerations"`
-		HTTP struct {
-			Timeout string `json:"timeout"`
-			RetryWaitMin string `json:"retryWaitMin"`
-			RetryWaitMax string `json:"retryWaitMax"`
-			RetryMax int `json:"retryMax"`
-			PanicOnFailure bool `json:"panicOnFailure"`
+		Enabled     bool                        `json:"enabled"`
+		Image       helm.Image                  `json:"image"`
+		Resources   corev1.ResourceRequirements `json:"resources"`
+		Tolerations []corev1.Toleration         `json:"tolerations"`
+		HTTP        struct {
+			Timeout        string `json:"timeout"`
+			RetryWaitMin   string `json:"retryWaitMin"`
+			RetryWaitMax   string `json:"retryWaitMax"`
+			RetryMax       int    `json:"retryMax"`
+			PanicOnFailure bool   `json:"panicOnFailure"`
 		} `json:"http"`
 	} `json:"auditsink"`
 
