@@ -12,15 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ts
+package cb
 
 import (
 	"emperror.dev/errors"
+	"github.com/AlecAivazis/survey/v2"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 
+	clierrors "github.com/banzaicloud/backyards-cli/internal/errors"
 	"github.com/banzaicloud/backyards-cli/pkg/cli"
 	"github.com/banzaicloud/backyards-cli/pkg/cli/cmd/routing/common"
 	"github.com/banzaicloud/backyards-cli/pkg/graphql"
@@ -44,7 +46,7 @@ func newDeleteCommand(cli cli.CLI) *cobra.Command {
 
 	cmd := &cobra.Command{
 		Use:           "delete [[--service=]namespace/servicename]",
-		Short:         "Delete traffic shifting rules of a service",
+		Short:         "Delete circuit breaker rules of a service",
 		Args:          cobra.MaximumNArgs(1),
 		SilenceErrors: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -84,28 +86,53 @@ func (c *deleteCommand) run(cli cli.CLI, options *deleteOptions) error {
 		return errors.WrapIf(err, "could not get service")
 	}
 
+	if cli.InteractiveTerminal() {
+		data, err := getCircuitBreakerRulesByServiceName(cli, options.serviceName)
+		if err != nil {
+			if clierrors.IsNotFound(err) {
+				log.Infof("no circuit breaker rules set for %s", options.serviceName)
+				return nil
+			}
+			return err
+		}
+
+		log.Info("current settings")
+
+		err = Output(cli, data)
+		if err != nil {
+			return err
+		}
+
+		confirmed := false
+		err = survey.AskOne(&survey.Confirm{Message: "Do you want to DELETE the circuit breaker rules?"}, &confirmed)
+		if err != nil {
+			return errors.WrapIf(err, "could not ask for confirmation")
+		}
+		if !confirmed {
+			return errors.New("deletion cancelled")
+		}
+	}
+
 	client, err := common.GetGraphQLClient(cli)
 	if err != nil {
 		return errors.WrapIf(err, "could not get initialized graphql client")
 	}
 
-	req := graphql.DisableHTTPRouteRequest{
+	req := graphql.DisableGlobalTrafficPolicyRequest{
 		Name:      service.Name,
 		Namespace: service.Namespace,
-		Rules: []string{
-			"Route",
-		},
+		Rules:     []string{"ConnectionPool", "OutlierDetection"},
 	}
-	r, err := client.DisableHTTPRoute(req)
+	r, err := client.DisableGlobalTrafficPolicy(req)
 	if err != nil {
 		return err
 	}
 
 	if !r {
-		return errors.New("unknown error: cannot delete traffic shifting")
+		return errors.New("unknown error: cannot delete circuit breaker rules")
 	}
 
-	log.Infof("traffic shifting rules set to %s successfully deleted", options.serviceName)
+	log.Infof("circuit breaker rules set to %s successfully deleted", options.serviceName)
 
 	return nil
 }
