@@ -15,18 +15,13 @@
 package ts
 
 import (
-	"context"
 	"fmt"
 	"regexp"
 	"sort"
 	"strconv"
 	"strings"
 
-	corev1 "k8s.io/api/core/v1"
 	"emperror.dev/errors"
-	"github.com/banzaicloud/backyards-cli/pkg/cli"
-	"k8s.io/apimachinery/pkg/types"
-	"knative.dev/pkg/apis/istio/v1alpha3"
 )
 
 const dns1123LabelFmt string = "[a-z0-9]([-a-z0-9]*[a-z0-9])?"
@@ -61,7 +56,7 @@ func (p parsedSubsets) Validate() error {
 }
 
 func parseSubsets(subsets []string) (parsedSubsets, error) {
-	parsedSubsets := make(parsedSubsets, 0)
+	parsedSubsets := make(parsedSubsets)
 
 	for _, subset := range subsets {
 		parts := strings.Split(subset, "=")
@@ -82,54 +77,4 @@ func parseSubsets(subsets []string) (parsedSubsets, error) {
 	}
 
 	return parsedSubsets, nil
-}
-
-func parseServiceID(serviceID string) (types.NamespacedName, error) {
-	parts := strings.Split(serviceID, "/")
-	if len(parts) != 2 {
-		return types.NamespacedName{}, errors.Errorf("invalid service ID: '%s': format must be <namespace>/<name>", serviceID)
-	}
-
-	for _, p := range parts {
-		if !dns1123LabelRegexp.MatchString(p) {
-			return types.NamespacedName{}, errors.Errorf("invalid service ID: '%s': format must be <namespace>/<name>", serviceID)
-		}
-	}
-
-	return types.NamespacedName{
-		Namespace: parts[0],
-		Name:      parts[1],
-	}, nil
-}
-
-func getVirtualserviceByName(cli cli.CLI, serviceName types.NamespacedName) (*v1alpha3.VirtualService, error) {
-	var vservice v1alpha3.VirtualService
-
-	k8sclient, err := cli.GetK8sClient()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	err = k8sclient.Get(context.Background(), serviceName, &vservice)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	return &vservice, nil
-}
-
-func getService(cli cli.CLI, serviceName types.NamespacedName) (*corev1.Service, error) {
-	var service corev1.Service
-
-	k8sclient, err := cli.GetK8sClient()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	err = k8sclient.Get(context.Background(), serviceName, &service)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	return &service, nil
 }
